pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "stockshark"
version = "0.1.0"
description = "Stock Analysis System - 股票分析系统"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "StockShark Team"}
]
keywords = ["stock", "analysis", "flask", "finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Flask",
]

dependencies = [
    "flask>=2.0.0",
    "flask-cors>=3.0.0",
    "pymysql>=1.0.0",
    "pymongo>=4.0.0",
    "akshare>=1.0.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "python-dotenv>=0.19.0",
    "jieba>=0.42.0",
    "requests>=2.26.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "isort>=5.9.0",
]
nlp = [
    "hanlp>=2.0.0",
    "spacy>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/stockshark"
Repository = "https://github.com/yourusername/stockshark.git"
Documentation = "https://github.com/yourusername/stockshark/blob/main/README.md"

[tool.setuptools]
packages = ["stockshark"]

[tool.setuptools_scm]
write_to = "stockshark/_version.py"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
